tests/wpt/web-platform-tests/bluetooth/characteristic/readValue/read-updates-value.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Succesful read should update characteristic.value';
const EXPECTED_VALUE = [0, 1, 2];
let characteristic, fake_characteristic;

bluetooth_test(() => getMeasurementIntervalCharacteristic()
    .then(_ => ({characteristic, fake_characteristic} = _))
    .then(() => assert_equals(characteristic.value, null))
    .then(() => fake_characteristic.setNextReadResponse(
          GATT_SUCCESS, EXPECTED_VALUE))
    .then(() => characteristic.readValue())
    .then(() => assert_array_equals(
          new Uint8Array(characteristic.value.buffer),
          EXPECTED_VALUE)),
    test_desc);
</script>
